+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [4.17.7](https://github.com/bharani-palani/ledgerely/compare/v4.17.6...v4.17.7) (2026-08-01)
+
+
+### Bug Fixes
+
+* datasource dynamic query vulnurability fix ([c989549](https://github.com/bharani-palani/ledgerely/commit/c989549e4ab7620aeb12ee5db3405503b52805bb))
+
+## [4.17.6](https://github.com/bharani-palani/ledgerely/compare/v4.17.5...v4.17.6) (2026-07-31)
+
+
+### Bug Fixes
+
+* Datasource, Dynamic clause inline edit added ([2d3b880](https://github.com/bharani-palani/ledgerely/commit/2d3b8800934e0cca729a59fb852755f128cab70b))
+
 ## [4.17.5](https://github.com/bharani-palani/ledgerely/compare/v4.17.4...v4.17.5) (2026-07-30)